+++ src/biz/bokhorst/xprivacy/XMediaRecorder.java
@@ -15,18 +15,24 @@ public class XMediaRecorder extends XHook {
 		return "android.media.MediaRecorder";
 	}
 
+	// void setOutputFile(FileDescriptor fd)
+	// void setOutputFile(String path)
+	// public prepare()
 	// public native void start()
+	// void stop()
 	// frameworks/base/media/java/android/media/MediaRecorder.java
 	// http://developer.android.com/reference/android/media/MediaRecorder.html
 
 	private enum Methods {
-		setOutputFile, start
+		setOutputFile, prepare, start, stop
 	};
 
 	public static List<XHook> getInstances() {
 		List<XHook> listHook = new ArrayList<XHook>();
 		listHook.add(new XMediaRecorder(Methods.setOutputFile, PrivacyManager.cMedia));
+		listHook.add(new XMediaRecorder(Methods.prepare, null));
 		listHook.add(new XMediaRecorder(Methods.start, PrivacyManager.cMedia));
+		listHook.add(new XMediaRecorder(Methods.stop, null));
 		return listHook;
 	}
 
@@ -36,6 +42,12 @@ public class XMediaRecorder extends XHook {
 		case setOutputFile:
 		case start:
 			if (isRestricted(param))
+				param.setResult(null);
+			break;
+
+		case prepare:
+		case stop:
+			if (isRestricted(param, PrivacyManager.cMedia, "MediaRecorder.start"))
 				param.setResult(null);
 			break;
 		}